chore: move `Webapp` resource to `webapp/resource` package
