+++ sample-operators/tomcat-operator/src/main/java/io/javaoperatorsdk/operator/sample/webapp/resource/Webapp.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.sample;
+package io.javaoperatorsdk.operator.sample.webapp.resource;
 
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
