+++ sample-operators/tomcat-operator/src/main/java/io/javaoperatorsdk/operator/sample/webapp/resource/WebappStatus.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.sample;
+package io.javaoperatorsdk.operator.sample.webapp.resource;
 
 public class WebappStatus {
